Move 'bundle' package out of 'machine'
It does not use anything machine specific, so we can move it outside of
'machine' to make it easier to reuse.
